Hide and prevent protected views and actions from the user that is not logged in

+++ src/components/moviesTable.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
 import { Link } from "react-router-dom";
+import auth from "../services/authService";
 import Like from "./common/like";
 import Table from "./common/table";
 class MoviesTable extends Component {
@@ -16,20 +17,27 @@ class MoviesTable extends Component {
     { path: "dailyRentalRate", label: "Rate" },
     {
       key: "like",
-      content: (movie) => (
-        <Like liked={movie.liked} onClick={() => this.props.onLike(movie)} />
-      ),
+      content: (movie) => {
+        if (!auth.getCurrentUser()) return null;
+        return (
+          <Like liked={movie.liked} onClick={() => this.props.onLike(movie)} />
+        );
+      },
     },
     {
       key: "delete",
-      content: (movie) => (
-        <button
-          onClick={() => this.props.onDelete(movie)}
-          className="btn btn-danger btn-sm"
-        >
-          Delete
-        </button>
-      ),
+      content: (movie) => {
+        const user = auth.getCurrentUser();
+        if (!(user && user.isAdmin)) return null;
+        return (
+          <button
+            onClick={() => this.props.onDelete(movie)}
+            className="btn btn-danger btn-sm"
+          >
+            Delete
+          </button>
+        );
+      },
     },
   ];
